requirements.txt
requests
python-dotenv
flask_sqlalchemy
flask_login
flask
python-dotenv
psycopg2